Rename deps
